requirements.txt

langchain
langchain-experimental
langchain-openai
qrcode
black
python-dotenv
langchainhub
pandas
tabulate
streamlit
